raft/serde config
