+++ src/v/raft/configuration.cc
@@ -13,6 +13,7 @@
 #include "model/metadata.h"
 #include "raft/consensus.h"
 #include "raft/consensus_utils.h"
+#include "raft/reflection.h"
 
 #include <absl/container/flat_hash_set.h>
 #include <bits/stdint-uintn.h>
