+++ src/v/raft/configuration_bootstrap_state.cc
@@ -11,7 +11,7 @@
 
 #include "likely.h"
 #include "model/record.h"
-#include "reflection/adl.h"
+#include "raft/reflection.h"
 
 #include <fmt/format.h>
