+++ src/v/raft/consensus_utils.cc
@@ -14,8 +14,8 @@
 #include "model/record.h"
 #include "model/timestamp.h"
 #include "raft/logger.h"
+#include "raft/reflection.h"
 #include "random/generators.h"
-#include "reflection/adl.h"
 #include "resource_mgmt/io_priority.h"
 #include "storage/record_batch_builder.h"
 #include "vassert.h"
