+++ src/v/raft/tests/foreign_entry_test.cc
@@ -16,7 +16,7 @@
 #include "raft/configuration.h"
 #include "raft/configuration_bootstrap_state.h"
 #include "raft/consensus_utils.h"
-#include "raft/types.h"
+#include "raft/reflection.h"
 #include "random/generators.h"
 #include "resource_mgmt/io_priority.h"
 #include "storage/api.h"
